#1297 Lint tests
